Incorporated error handles and periodic BPM
incorporated a module for asserting that the data types int the input module are correct. Also wrote a module to write out periodic BPM data files. This is still not working as of 8/11. The forward flight loading module has been modified so that the desired and computed integrated thrust coefficient now match.