Fixed TE thickness calc and added the OLS example case
Corrected the calculation for the TE thickness that is used in the BPM broadband model. The TE thickness is determined by determining the difference in the Z direction of the points two tessellations away from the TE. An additional example case of the OLS rotor in forward flight was also added to the test case folder.